Continued to work on the tagging pipeline
