+++ dwh/etl/aggregation/tagging pipeline/postprocess.py
@@ -1,0 +1,9 @@
+"""
+This module contains the postprocessing logic for the tagging pipeline.
+"""
+
+
+def postprocess_data(df):
+    # TODO: Implement postprocessing logic
+    print("Postprocessing data...")
+    return df
